added knapsack
